Update dependencies.
This addresses some critical audit problems. Also the upgrade to XLSX allows it to work in React. There are a few dev dependencies that fail audit, namely istanbul, which is a deprecated package. Perhaps in a future PR this can be replaced with nyc (https://www.npmjs.com/package/nyc)